De-emphasizes the MessageTable class, by making it a private module in the recipe parser project and changing documentation
