+++ conda_recipe_manager/commands/convert.py
@@ -16,12 +16,12 @@ from typing import Final, Optional
 
 import click
 
-from conda_recipe_manager.commands.utils.print import print_err, print_messages, print_out
+from conda_recipe_manager.commands.utils.print import print_err, print_out
 from conda_recipe_manager.commands.utils.types import CONTEXT_SETTINGS, ExitCode
+from conda_recipe_manager.parser._message_table import MessageCategory, MessageTable
 from conda_recipe_manager.parser.enums import SchemaVersion
 from conda_recipe_manager.parser.recipe_parser_convert import RecipeParserConvert
 from conda_recipe_manager.parser.types import V0_FORMAT_RECIPE_FILE_NAME, V1_FORMAT_RECIPE_FILE_NAME
-from conda_recipe_manager.types import MessageCategory, MessageTable
 
 # When performing a bulk operation, overall "success" is indicated by the % of recipe files that were converted
 # "successfully"
@@ -292,8 +292,8 @@ def convert(
     ## Single-file case ##
     if len(files) == 1:
         result: Final[ConversionResult] = convert_file(files[0], output, True, debug)
-        print_messages(MessageCategory.WARNING, result.msg_tbl)
-        print_messages(MessageCategory.ERROR, result.msg_tbl)
+        result.msg_tbl.print_messages_by_category(MessageCategory.WARNING)
+        result.msg_tbl.print_messages_by_category(MessageCategory.ERROR)
         print_err(result.msg_tbl.get_totals_message())
         sys.exit(result.code)
